pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager"
authors = ["Elizaveta"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
python-dotenv = "^1.0.1"
django = "^5.0.2"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
uvicorn = "^0.27.1"
django-bootstrap5 = "^23.4"
django-filter = "^23.5"
rollbar = "^0.16.3"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.8.0"
flake8 = "^7.0.0"
coverage = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
